BIM: Structure: Refactor: Separated command from object
