+++ src/Mod/BIM/ArchStructure.py
@@ -20,26 +20,15 @@
 #***************************************************************************
 #Modified 2016-01-03 JAndersM
 
-import FreeCAD,Draft,ArchComponent,DraftVecUtils,ArchCommands
+import FreeCAD
+import Draft
+import ArchComponent
+import DraftVecUtils
+import ArchCommands
 from FreeCAD import Vector
 import ArchProfile
 from draftutils import params
-from draftutils import gui_utils
 
-if FreeCAD.GuiUp:
-    import FreeCADGui
-    from PySide import QtCore, QtGui
-    from draftutils.translate import translate
-    from PySide.QtCore import QT_TRANSLATE_NOOP
-    import ArchPrecast
-    import draftguitools.gui_trackers as DraftTrackers
-else:
-    # \cond
-    def translate(ctxt,txt):
-        return txt
-    def QT_TRANSLATE_NOOP(ctxt,txt):
-        return txt
-    # \endcond
 
 ## @package ArchStructure
 #  \ingroup ARCH
@@ -63,611 +52,7 @@ for pre in Presets:
         Categories.append(pre[1])
 
 
-def makeStructure(baseobj=None,length=None,width=None,height=None,name=None):
-
-    '''makeStructure([baseobj],[length],[width],[height],[name]): creates a
-    structure element based on the given profile object and the given
-    extrusion height. If no base object is given, you can also specify
-    length and width for a cubic object.'''
-
-    if not FreeCAD.ActiveDocument:
-        FreeCAD.Console.PrintError("No active document. Aborting\n")
-        return
-    obj = FreeCAD.ActiveDocument.addObject("Part::FeaturePython","Structure")
-    _Structure(obj)
-    if FreeCAD.GuiUp:
-        _ViewProviderStructure(obj.ViewObject)
-    if baseobj:
-        obj.Base = baseobj
-        if FreeCAD.GuiUp:
-            obj.Base.ViewObject.hide()
-    if width:
-        obj.Width = width
-    else:
-        obj.Width = params.get_param_arch("StructureWidth")
-    if height:
-        obj.Height = height
-    else:
-        if not length:
-            obj.Height = params.get_param_arch("StructureHeight")
-    if length:
-        obj.Length = length
-    else:
-        if not baseobj:
-            # don't set the length if we have a base object, otherwise the length X height calc
-            # gets wrong
-            obj.Length = params.get_param_arch("StructureLength")
-    if baseobj:
-        w = 0
-        h = 0
-        if hasattr(baseobj,"Width") and hasattr(baseobj,"Height"):
-            w = baseobj.Width.Value
-            h = baseobj.Height.Value
-        elif hasattr(baseobj,"Length") and hasattr(baseobj,"Width"):
-            w = baseobj.Length.Value
-            h = baseobj.Width.Value
-        elif hasattr(baseobj,"Length") and hasattr(baseobj,"Height"):
-            w = baseobj.Length.Value
-            h = baseobj.Height.Value
-        if w and h:
-            if length and not height:
-                obj.Width = w
-                obj.Height = h
-            elif height and not length:
-                obj.Width = w
-                obj.Length = h
-
-    if not height and not length:
-        obj.IfcType = "Building Element Proxy"
-        obj.Label = name if name else translate("Arch","Structure")
-    elif obj.Length > obj.Height:
-        obj.IfcType = "Beam"
-        obj.Label = name if name else translate("Arch","Beam")
-    elif obj.Height > obj.Length:
-        obj.IfcType = "Column"
-        obj.Label = name if name else translate("Arch","Column")
-    return obj
-
-def makeStructuralSystem(objects=[],axes=[],name=None):
-
-    '''makeStructuralSystem([objects],[axes],[name]): makes a structural system
-    based on the given objects and axes'''
-
-    if not FreeCAD.ActiveDocument:
-        FreeCAD.Console.PrintError("No active document. Aborting\n")
-        return
-    result = []
-    if not axes:
-        print("At least one axis must be given")
-        return
-    if objects:
-        if not isinstance(objects,list):
-            objects = [objects]
-    else:
-        objects = [None]
-    for o in objects:
-        obj = FreeCAD.ActiveDocument.addObject("Part::FeaturePython","StructuralSystem")
-        obj.Label = name if name else translate("Arch","StructuralSystem")
-        _StructuralSystem(obj)
-        if FreeCAD.GuiUp:
-            _ViewProviderStructuralSystem(obj.ViewObject)
-        if o:
-            obj.Base = o
-        obj.Axes = axes
-        result.append(obj)
-        if FreeCAD.GuiUp and o:
-            o.ViewObject.hide()
-            Draft.formatObject(obj,o)
-    FreeCAD.ActiveDocument.recompute()
-    if len(result) == 1:
-        return result[0]
-    else:
-        return result
-
-def placeAlongEdge(p1,p2,horizontal=False):
-
-    """placeAlongEdge(p1,p2,[horizontal]): returns a Placement positioned at p1, with Z axis oriented towards p2.
-    If horizontal is True, then the X axis is oriented towards p2, not the Z axis"""
-
-    pl = FreeCAD.Placement()
-    pl.Base = p1
-    import WorkingPlane
-    up = WorkingPlane.get_working_plane(update=False).axis
-    zaxis = p2.sub(p1)
-    yaxis = up.cross(zaxis)
-    if yaxis.Length > 0:
-        xaxis = zaxis.cross(yaxis)
-        if horizontal:
-            pl.Rotation = FreeCAD.Rotation(zaxis,yaxis,xaxis,"ZXY")
-        else:
-            pl.Rotation = FreeCAD.Rotation(xaxis,yaxis,zaxis,"ZXY")
-            pl.Rotation = FreeCAD.Rotation(pl.Rotation.multVec(FreeCAD.Vector(0,0,1)),90).multiply(pl.Rotation)
-    return pl
-
-
-class CommandStructuresFromSelection:
-    """ The Arch Structures from selection command definition. """
-
-    def __init__(self):
-        pass
-
-    def GetResources(self):
-        return {'Pixmap': 'Arch_MultipleStructures',
-                'MenuText': QT_TRANSLATE_NOOP("Arch_StructuresFromSelection", "Multiple Structures"),
-                'ToolTip': QT_TRANSLATE_NOOP("Arch_StructuresFromSelection", "Create multiple BIM Structures from a selected base, using each selected edge as an extrusion path")}
-
-    def IsActive(self):
-        v = hasattr(FreeCADGui.getMainWindow().getActiveWindow(), "getSceneGraph")
-        return v
-
-    def Activated(self):
-        selex = FreeCADGui.Selection.getSelectionEx()
-        if len(selex) >= 2:
-            FreeCAD.ActiveDocument.openTransaction(translate("Arch", "Create Structures From Selection"))
-            FreeCADGui.addModule("Arch")
-            FreeCADGui.addModule("Draft")
-            base = selex[0].Object # The first selected object is the base for the Structure objects
-            for selexi in selex[1:]: # All the edges from the other objects are used as a Tool (extrusion paths)
-                if len(selexi.SubElementNames) == 0:
-                    subelement_names = ["Edge" + str(i) for i in range(1, len(selexi.Object.Shape.Edges) + 1)]
-                else:
-                    subelement_names = [sub for sub in selexi.SubElementNames if sub.startswith("Edge")]
-                for sub in subelement_names:
-                    FreeCADGui.doCommand("structure = Arch.makeStructure(FreeCAD.ActiveDocument." + base.Name + ")")
-                    FreeCADGui.doCommand("structure.Tool = (FreeCAD.ActiveDocument." + selexi.Object.Name + ", '" + sub + "')")
-                    FreeCADGui.doCommand("structure.BasePerpendicularToTool = True")
-                    FreeCADGui.doCommand("Draft.autogroup(structure)")
-            FreeCAD.ActiveDocument.commitTransaction()
-            FreeCAD.ActiveDocument.recompute()
-        else:
-            FreeCAD.Console.PrintError(translate("Arch", "Please select the base object first and then the edges to use as extrusion paths") + "\n")
-
-
-class CommandStructuralSystem:
-    """ The Arch Structural System command definition. """
-
-    def __init__(self):
-        pass
-
-    def GetResources(self):
-        return {'Pixmap': 'Arch_StructuralSystem',
-                'MenuText': QT_TRANSLATE_NOOP("Arch_StructuralSystem", "Structural System"),
-                'ToolTip': QT_TRANSLATE_NOOP("Arch_StructuralSystem", "Create a structural system from a selected structure and axis")}
-
-    def IsActive(self):
-        return not FreeCAD.ActiveDocument is None
-
-    def Activated(self):
-        sel = FreeCADGui.Selection.getSelection()
-        if sel:
-            st = Draft.getObjectsOfType(sel, "Structure")
-            ax = Draft.getObjectsOfType(sel, "Axis")
-            if ax:
-                FreeCAD.ActiveDocument.openTransaction(translate("Arch", "Create Structural System"))
-                FreeCADGui.addModule("Arch")
-                if st:
-                    FreeCADGui.doCommand("obj = Arch.makeStructuralSystem(" + ArchCommands.getStringList(st) + ", " + ArchCommands.getStringList(ax) + ")")
-                else:
-                    FreeCADGui.doCommand("obj = Arch.makeStructuralSystem(axes = " + ArchCommands.getStringList(ax) + ")")
-                FreeCADGui.addModule("Draft")
-                FreeCADGui.doCommand("Draft.autogroup(obj)")
-                FreeCAD.ActiveDocument.commitTransaction()
-                FreeCAD.ActiveDocument.recompute()
-            else:
-                FreeCAD.Console.PrintError(translate("Arch", "Please select at least an axis object") + "\n")
-
-
-class _CommandStructure:
-
-    "the Arch Structure command definition"
-
-    def __init__(self):
-
-        self.beammode = False
-
-    def GetResources(self):
-
-        return {'Pixmap'  : 'Arch_Structure',
-                'MenuText': QT_TRANSLATE_NOOP("Arch_Structure","Structure"),
-                'Accel': "S, T",
-                'ToolTip': QT_TRANSLATE_NOOP("Arch_Structure","Creates a structure from scratch or from a selected object (sketch, wire, face or solid)")}
-
-    def IsActive(self):
-
-        return not FreeCAD.ActiveDocument is None
-
-    def Activated(self):
-
-        self.Width = params.get_param_arch("StructureWidth")
-        if self.beammode:
-            self.Height = params.get_param_arch("StructureLength")
-            self.Length = params.get_param_arch("StructureHeight")
-        else:
-            self.Length = params.get_param_arch("StructureLength")
-            self.Height = params.get_param_arch("StructureHeight")
-        self.Profile = None
-        self.continueCmd = False
-        self.bpoint = None
-        self.bmode = False
-        self.precastvalues = None
-        self.wp = None
-        sel = FreeCADGui.Selection.getSelection()
-        if sel:
-            st = Draft.getObjectsOfType(sel,"Structure")
-            ax = Draft.getObjectsOfType(sel,"Axis")
-            if ax:
-                FreeCADGui.runCommand("Arch_StructuralSystem")
-                return
-            elif not(ax) and not(st):
-                FreeCAD.ActiveDocument.openTransaction(translate("Arch","Create Structure"))
-                FreeCADGui.addModule("Arch")
-                for obj in sel:
-                    FreeCADGui.doCommand("obj = Arch.makeStructure(FreeCAD.ActiveDocument." + obj.Name + ")")
-                    FreeCADGui.addModule("Draft")
-                    FreeCADGui.doCommand("Draft.autogroup(obj)")
-                FreeCAD.ActiveDocument.commitTransaction()
-                FreeCAD.ActiveDocument.recompute()
-                return
-
-        # interactive mode
-        import WorkingPlane
-        self.wp = WorkingPlane.get_working_plane()
-
-        self.points = []
-        self.tracker = DraftTrackers.boxTracker()
-        self.tracker.width(self.Width)
-        self.tracker.height(self.Height)
-        self.tracker.length(self.Length)
-        self.tracker.setRotation(self.wp.get_placement().Rotation)
-        self.tracker.on()
-        self.precast = ArchPrecast._PrecastTaskPanel()
-        self.dents = ArchPrecast._DentsTaskPanel()
-        self.precast.Dents = self.dents
-        if self.beammode:
-            title=translate("Arch","First point of the beam")+":"
-        else:
-            title=translate("Arch","Base point of column")+":"
-        FreeCAD.activeDraftCommand = self  # register as a Draft command for auto grid on/off
-        FreeCADGui.Snapper.getPoint(callback=self.getPoint,movecallback=self.update,extradlg=[self.taskbox(),self.precast.form,self.dents.form],title=title)
-
-    def getPoint(self,point=None,obj=None):
-
-        "this function is called by the snapper when it has a 3D point"
-
-        self.bmode = self.modeb.isChecked()
-        if point is None:
-            self.tracker.finalize()
-            FreeCAD.activeDraftCommand = None
-            FreeCADGui.Snapper.off()
-            return
-        if self.bmode and (self.bpoint is None):
-            self.bpoint = point
-            FreeCADGui.Snapper.getPoint(last=point,callback=self.getPoint,movecallback=self.update,extradlg=[self.taskbox(),self.precast.form,self.dents.form],title=translate("Arch","Next point")+":",mode="line")
-            return
-        self.tracker.off()
-        FreeCAD.activeDraftCommand = None
-        FreeCADGui.Snapper.off()
-        horiz = True # determines the type of rotation to apply to the final object
-        FreeCAD.ActiveDocument.openTransaction(translate("Arch","Create Structure"))
-        FreeCADGui.addModule("Arch")
-        FreeCADGui.addModule("WorkingPlane")
-        if self.bmode:
-            self.Length = point.sub(self.bpoint).Length
-            params.set_param_arch("StructureHeight",self.Length)
-        if self.Profile is not None:
-            try: # try to update latest precast values - fails if dialog has been destroyed already
-                self.precastvalues = self.precast.getValues()
-            except Exception:
-                pass
-            if ("Precast" in self.Profile) and self.precastvalues:
-                # precast concrete
-                self.precastvalues["PrecastType"] = self.Profile.split("_")[1]
-                self.precastvalues["Length"] = self.Length
-                self.precastvalues["Width"] = self.Width
-                self.precastvalues["Height"] = self.Height
-                argstring = ""
-                # fix for precast placement, since their (0,0) point is the lower left corner
-                if self.bmode:
-                    delta = FreeCAD.Vector(0,0-self.Width/2,0)
-                else:
-                    delta = FreeCAD.Vector(-self.Length/2,-self.Width/2,0)
-                delta = self.wp.get_global_coords(delta,as_vector=True)
-                point = point.add(delta)
-                if self.bpoint:
-                    self.bpoint = self.bpoint.add(delta)
-                # build the string definition
-                for pair in self.precastvalues.items():
-                    argstring += pair[0].lower() + "="
-                    if isinstance(pair[1],str):
-                        argstring += '"' + pair[1] + '",'
-                    else:
-                        argstring += str(pair[1]) + ","
-                FreeCADGui.addModule("ArchPrecast")
-                FreeCADGui.doCommand("s = ArchPrecast.makePrecast("+argstring+")")
-            else:
-                # metal profile
-                FreeCADGui.doCommand('p = Arch.makeProfile('+str(self.Profile)+')')
-                if (abs(self.Length - self.Profile[4]) >= 0.1) or self.bmode: # forgive rounding errors
-                    # horizontal
-                    FreeCADGui.doCommand('s = Arch.makeStructure(p,length='+str(self.Length)+')')
-                    horiz = False
-                else:
-                    # vertical
-                    FreeCADGui.doCommand('s = Arch.makeStructure(p,height='+str(self.Height)+')')
-                    #if not self.bmode:
-                    #    FreeCADGui.doCommand('s.Placement.Rotation = FreeCAD.Rotation(-0.5,0.5,-0.5,0.5)')
-                FreeCADGui.doCommand('s.Profile = "'+self.Profile[2]+'"')
-        else :
-            FreeCADGui.doCommand('s = Arch.makeStructure(length='+str(self.Length)+',width='+str(self.Width)+',height='+str(self.Height)+')')
-
-        # calculate rotation
-        if self.bmode and self.bpoint:
-            FreeCADGui.doCommand('s.Placement = Arch.placeAlongEdge('+DraftVecUtils.toString(self.bpoint)+","+DraftVecUtils.toString(point)+","+str(horiz)+")")
-        else:
-            FreeCADGui.doCommand('s.Placement.Base = '+DraftVecUtils.toString(point))
-            FreeCADGui.doCommand('wp = WorkingPlane.get_working_plane()')
-            FreeCADGui.doCommand('s.Placement.Rotation = s.Placement.Rotation.multiply(wp.get_placement().Rotation)')
-
-        FreeCADGui.addModule("Draft")
-        FreeCADGui.doCommand("Draft.autogroup(s)")
-        FreeCAD.ActiveDocument.commitTransaction()
-        FreeCAD.ActiveDocument.recompute()
-        # gui_utils.end_all_events()  # Causes a crash on Linux.
-        self.tracker.finalize()
-        if self.continueCmd:
-            self.Activated()
-
-    def _createItemlist(self, baselist):
-
-        "create nice labels for presets in the task panel"
-
-        ilist=[]
-        for p in baselist:
-            f = FreeCAD.Units.Quantity(p[4],FreeCAD.Units.Length).getUserPreferred()
-            d = params.get_param("Decimals",path="Units")
-            s1 = str(round(p[4]/f[1],d))
-            s2 = str(round(p[5]/f[1],d))
-            s3 = str(f[2])
-            ilist.append(p[2]+" ("+s1+"x"+s2+s3+")")
-        return ilist
-
-    def taskbox(self):
-
-        "sets up a taskbox widget"
-
-        w = QtGui.QWidget()
-        ui = FreeCADGui.UiLoader()
-        w.setWindowTitle(translate("Arch","Structure options"))
-        grid = QtGui.QGridLayout(w)
-
-        # mode box
-        labelmode = QtGui.QLabel(translate("Arch","Parameters of the structure")+":")
-        self.modeb = QtGui.QRadioButton(translate("Arch","Beam"))
-        self.modec = QtGui.QRadioButton(translate("Arch","Column"))
-        if self.bpoint or self.beammode:
-            self.modeb.setChecked(True)
-        else:
-            self.modec.setChecked(True)
-        grid.addWidget(labelmode,0,0,1,2)
-        grid.addWidget(self.modeb,1,0,1,1)
-        grid.addWidget(self.modec,1,1,1,1)
-
-        # categories box
-        labelc = QtGui.QLabel(translate("Arch","Category"))
-        self.valuec = QtGui.QComboBox()
-        self.valuec.addItems([" ","Precast concrete"]+Categories)
-        grid.addWidget(labelc,2,0,1,1)
-        grid.addWidget(self.valuec,2,1,1,1)
-
-        # presets box
-        labelp = QtGui.QLabel(translate("Arch","Preset"))
-        self.vPresets = QtGui.QComboBox()
-        self.pSelect = [None]
-        fpresets = [" "]
-        self.vPresets.addItems(fpresets)
-        grid.addWidget(labelp,3,0,1,1)
-        grid.addWidget(self.vPresets,3,1,1,1)
-
-        # length
-        label1 = QtGui.QLabel(translate("Arch","Length"))
-        self.vLength = ui.createWidget("Gui::InputField")
-        if self.modeb.isChecked():
-            self.vLength.setText(FreeCAD.Units.Quantity(self.Height,FreeCAD.Units.Length).UserString)
-        else:
-            self.vLength.setText(FreeCAD.Units.Quantity(self.Length,FreeCAD.Units.Length).UserString)
-        grid.addWidget(label1,4,0,1,1)
-        grid.addWidget(self.vLength,4,1,1,1)
-
-        # width
-        label2 = QtGui.QLabel(translate("Arch","Width"))
-        self.vWidth = ui.createWidget("Gui::InputField")
-        self.vWidth.setText(FreeCAD.Units.Quantity(self.Width,FreeCAD.Units.Length).UserString)
-        grid.addWidget(label2,5,0,1,1)
-        grid.addWidget(self.vWidth,5,1,1,1)
-
-        # height
-        label3 = QtGui.QLabel(translate("Arch","Height"))
-        self.vHeight = ui.createWidget("Gui::InputField")
-        if self.modeb.isChecked():
-            self.vHeight.setText(FreeCAD.Units.Quantity(self.Length,FreeCAD.Units.Length).UserString)
-        else:
-            self.vHeight.setText(FreeCAD.Units.Quantity(self.Height,FreeCAD.Units.Length).UserString)
-        grid.addWidget(label3,6,0,1,1)
-        grid.addWidget(self.vHeight,6,1,1,1)
-
-        # horizontal button
-        value5 = QtGui.QPushButton(translate("Arch","Switch Length/Height"))
-        grid.addWidget(value5,7,0,1,1)
-        value6 = QtGui.QPushButton(translate("Arch","Switch Length/Width"))
-        grid.addWidget(value6,7,1,1,1)
-
-        # continue button
-        label4 = QtGui.QLabel(translate("Arch","Con&tinue"))
-        value4 = QtGui.QCheckBox()
-        value4.setObjectName("ContinueCmd")
-        value4.setLayoutDirection(QtCore.Qt.RightToLeft)
-        label4.setBuddy(value4)
-        self.continueCmd = params.get_param("ContinueMode")
-        value4.setChecked(self.continueCmd)
-        grid.addWidget(label4,8,0,1,1)
-        grid.addWidget(value4,8,1,1,1)
-
-        # connect slots
-        QtCore.QObject.connect(self.valuec,QtCore.SIGNAL("currentIndexChanged(int)"),self.setCategory)
-        QtCore.QObject.connect(self.vPresets,QtCore.SIGNAL("currentIndexChanged(int)"),self.setPreset)
-        QtCore.QObject.connect(self.vLength,QtCore.SIGNAL("valueChanged(double)"),self.setLength)
-        QtCore.QObject.connect(self.vWidth,QtCore.SIGNAL("valueChanged(double)"),self.setWidth)
-        QtCore.QObject.connect(self.vHeight,QtCore.SIGNAL("valueChanged(double)"),self.setHeight)
-        QtCore.QObject.connect(value4,QtCore.SIGNAL("stateChanged(int)"),self.setContinue)
-        QtCore.QObject.connect(value5,QtCore.SIGNAL("pressed()"),self.rotateLH)
-        QtCore.QObject.connect(value6,QtCore.SIGNAL("pressed()"),self.rotateLW)
-        QtCore.QObject.connect(self.modeb,QtCore.SIGNAL("toggled(bool)"),self.switchLH)
-
-        # restore preset
-        stored = params.get_param_arch("StructurePreset")
-        if stored:
-            if stored.lower().startswith("precast_"):
-                self.valuec.setCurrentIndex(1)
-                tp = stored.split("_")[1]
-                if tp and (tp in self.precast.PrecastTypes):
-                    self.vPresets.setCurrentIndex(self.precast.PrecastTypes.index(tp))
-            elif ";" in stored:
-                stored = stored.split(";")
-                if len(stored) >= 3:
-                    if stored[1] in Categories:
-                        self.valuec.setCurrentIndex(2+Categories.index(stored[1]))
-                        ps = [p[2] for p in self.pSelect]
-                        if stored[2] in ps:
-                            self.vPresets.setCurrentIndex(ps.index(stored[2]))
-        return w
-
-    def update(self,point,info):
-
-        "this function is called by the Snapper when the mouse is moved"
-
-        if FreeCADGui.Control.activeDialog():
-            try: # try to update latest precast values - fails if dialog has been destroyed already
-                self.precastvalues = self.precast.getValues()
-            except Exception:
-                pass
-            if self.Height >= self.Length:
-                delta = Vector(0,0,self.Height/2)
-            else:
-                delta = Vector(self.Length/2,0,0)
-            delta = self.wp.get_global_coords(delta,as_vector=True)
-            if self.modec.isChecked():
-                self.tracker.pos(point.add(delta))
-                self.tracker.on()
-            else:
-                if self.bpoint:
-                    delta = Vector(0,0,-self.Height/2)
-                    delta = self.wp.get_global_coords(delta,as_vector=True)
-                    self.tracker.update([self.bpoint.add(delta),point.add(delta)])
-                    self.tracker.on()
-                    l = (point.sub(self.bpoint)).Length
-                    self.vLength.setText(FreeCAD.Units.Quantity(l,FreeCAD.Units.Length).UserString)
-                else:
-                    self.tracker.off()
-
-    def setWidth(self,d):
-
-        self.Width = d
-        self.tracker.width(d)
-        params.set_param_arch("StructureWidth",d)
-
-    def setHeight(self,d):
-
-        self.Height = d
-        self.tracker.height(d)
-        if self.modeb.isChecked():
-            params.set_param_arch("StructureLength",d)
-        else:
-            params.set_param_arch("StructureHeight",d)
-
-    def setLength(self,d):
-
-        self.Length = d
-        self.tracker.length(d)
-        if self.modeb.isChecked():
-            params.set_param_arch("StructureHeight",d)
-        else:
-            params.set_param_arch("StructureLength",d)
-
-    def setContinue(self,i):
-
-        self.continueCmd = bool(i)
-        params.set_param("ContinueMode", bool(i))
-
-    def setCategory(self,i):
-
-        self.vPresets.clear()
-        if i > 1:
-            self.precast.form.hide()
-            self.pSelect = [p for p in Presets if p[1] == Categories[i-2]]
-            fpresets = self._createItemlist(self.pSelect)
-            self.vPresets.addItems(fpresets)
-            self.setPreset(0)
-        elif i == 1:
-            self.precast.form.show()
-            self.pSelect = self.precast.PrecastTypes
-            fpresets = self.precast.PrecastTypes
-            self.vPresets.addItems(fpresets)
-            self.setPreset(0)
-        else:
-            self.precast.form.hide()
-            self.pSelect = [None]
-            fpresets = [" "]
-            self.vPresets.addItems(fpresets)
-            params.set_param_arch("StructurePreset","")
-
-    def setPreset(self,i):
-
-        self.Profile = None
-        elt = self.pSelect[i]
-        if elt:
-            if elt in self.precast.PrecastTypes:
-                self.precast.setPreset(elt)
-                self.Profile = "Precast_" + elt
-                if elt in ["Pillar","Beam"]:
-                    self.dents.form.show()
-                else:
-                    self.dents.form.hide()
-                params.set_param_arch("StructurePreset",self.Profile)
-            else:
-                p=elt[0]-1 # Presets indexes are 1-based
-                self.vLength.setText(FreeCAD.Units.Quantity(float(Presets[p][4]),FreeCAD.Units.Length).UserString)
-                self.vWidth.setText(FreeCAD.Units.Quantity(float(Presets[p][5]),FreeCAD.Units.Length).UserString)
-                self.Profile = Presets[p]
-                params.set_param_arch("StructurePreset",";".join([str(i) for i in self.Profile]))
-
-    def switchLH(self,bmode):
-
-        if bmode:
-            self.bmode = True
-            if self.Height > self.Length:
-                self.rotateLH()
-        else:
-            self.bmode = False
-            if self.Length > self.Height:
-                self.rotateLH()
-                self.tracker.setRotation(FreeCAD.Rotation())
-
-    def rotateLH(self):
-
-        l = self.vLength.text()
-        h = self.vHeight.text()
-        self.vLength.setText(h)
-        self.vHeight.setText(l)
-
-    def rotateLW(self):
-
-        l = self.vLength.text()
-        w = self.vWidth.text()
-        self.vLength.setText(w)
-        self.vWidth.setText(l)
-
-
-class _Structure(ArchComponent.Component):
+class Structure(ArchComponent.Component):
 
     "The Structure object"
 
@@ -1173,7 +558,7 @@ class _Structure(ArchComponent.Component):
         return edges
 
 
-class _ViewProviderStructure(ArchComponent.ViewProviderComponent):
+class ViewProviderStructure(ArchComponent.ViewProviderComponent):
 
     "A View Provider for the Structure object"
 
@@ -1529,130 +914,6 @@ class StructSelectionObserver:
         self.callback(obj)
 
 
-class _StructuralSystem(ArchComponent.Component): # OBSOLETE - All Arch objects can now be based on axes
-
-    "The Structural System object"
-
-    def __init__(self,obj):
-
-        ArchComponent.Component.__init__(self,obj)
-        obj.addProperty("App::PropertyLinkList","Axes","Arch",QT_TRANSLATE_NOOP("App::Property","Axes systems this structure is built on"))
-        obj.addProperty("App::PropertyIntegerList","Exclude","Arch",QT_TRANSLATE_NOOP("App::Property","The element numbers to exclude when this structure is based on axes"))
-        obj.addProperty("App::PropertyBool","Align","Arch",QT_TRANSLATE_NOOP("App::Property","If true the element are aligned with axes")).Align = False
-        self.Type = "StructuralSystem"
-
-    def execute(self,obj):
-        "creates the structure shape"
-
-        import Part, DraftGeomUtils
-
-        # creating base shape
-        pl = obj.Placement
-        if obj.Base:
-            if hasattr(obj.Base,'Shape'):
-                if obj.Base.Shape.isNull():
-                    return
-                if not obj.Base.Shape.Solids:
-                    return
-
-                base = None
-
-                # applying axes
-                pts = self.getAxisPoints(obj)
-                if hasattr(obj,"Align"):
-                    if obj.Align == False :
-                        apl = self.getAxisPlacement(obj)
-                    if obj.Align:
-                        apl = None
-                else :
-                    apl = self.getAxisPlacement(obj)
-
-                if pts:
-                    fsh = []
-                    for i in range(len(pts)):
-                        sh = obj.Base.Shape.copy()
-                        if hasattr(obj,"Exclude"):
-                            if i in obj.Exclude:
-                                continue
-                        if apl:
-                            sh.Placement.Rotation = sh.Placement.Rotation.multiply(apl.Rotation)
-                        sh.translate(pts[i])
-                        fsh.append(sh)
-
-                    if fsh:
-                        base = Part.makeCompound(fsh)
-                        base = self.processSubShapes(obj,base,pl)
-
-                if base:
-                    if not base.isNull():
-                        if base.isValid() and base.Solids:
-                            if base.Volume < 0:
-                                base.reverse()
-                            if base.Volume < 0:
-                                FreeCAD.Console.PrintError(translate("Arch","Couldn't compute a shape"))
-                                return
-                            base = base.removeSplitter()
-                            obj.Shape = base
-                            if not pl.isNull():
-                                obj.Placement = pl
-
-    def getAxisPoints(self,obj):
-        "returns the gridpoints of linked axes"
-        import DraftGeomUtils
-        pts = []
-        if len(obj.Axes) == 1:
-            if hasattr(obj,"Align"):
-                if obj.Align:
-                    p0 = obj.Axes[0].Shape.Edges[0].Vertexes[1].Point
-                    for e in obj.Axes[0].Shape.Edges:
-                        p = e.Vertexes[1].Point
-                        p = p.sub(p0)
-                        pts.append(p)
-                else:
-                    for e in obj.Axes[0].Shape.Edges:
-                        pts.append(e.Vertexes[0].Point)
-            else:
-                for e in obj.Axes[0].Shape.Edges:
-                        pts.append(e.Vertexes[0].Point)
-        elif len(obj.Axes) >= 2:
-            set1 = obj.Axes[0].Shape.Edges
-            set2 = obj.Axes[1].Shape.Edges
-            for e1 in set1:
-                for e2 in set2:
-                    pts.extend(DraftGeomUtils.findIntersection(e1,e2))
-        return pts
-
-    def getAxisPlacement(self,obj):
-        "returns an axis placement"
-        if obj.Axes:
-            return obj.Axes[0].Placement
-        return None
-
-
-class _ViewProviderStructuralSystem(ArchComponent.ViewProviderComponent):
-
-    "A View Provider for the Structural System object"
-
-    def getIcon(self):
-
-        import Arch_rc
-        return ":/icons/Arch_StructuralSystem_Tree.svg"
-
-
-if FreeCAD.GuiUp:
-    FreeCADGui.addCommand("Arch_Structure", _CommandStructure())
-    FreeCADGui.addCommand("Arch_StructuralSystem", CommandStructuralSystem())
-    FreeCADGui.addCommand("Arch_StructuresFromSelection", CommandStructuresFromSelection())
-
-    class _ArchStructureGroupCommand:
-
-        def GetCommands(self):
-            return ("Arch_Structure", "Arch_StructuralSystem", "Arch_StructuresFromSelection")
-        def GetResources(self):
-            return { "MenuText": QT_TRANSLATE_NOOP("Arch_StructureTools", "Structure tools"),
-                     "ToolTip": QT_TRANSLATE_NOOP("Arch_StructureTools", "Structure tools")
-                   }
-        def IsActive(self):
-            return not FreeCAD.ActiveDocument is None
-
-    FreeCADGui.addCommand("Arch_StructureTools", _ArchStructureGroupCommand())
+# backwards compatibility
+_Structure = Structure
+_ViewProviderStructure = ViewProviderStructure
